Don't break existing builds, clean up method
